split qwen36 moe mtp loading
